+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 2.27.1-0 / Unreleased
-  *
+  * #819: ext-filter: Handle nodes without title
 
 # 2.27.0 / 2017-12-16
   * **BREAKING CHANGES:**